command error improvements migrate the embed over to components v2 and make the missing required argument error more descriptive
