+++ events/on_command_error.py
@@ -136,20 +136,7 @@ class OnCommandError(commands.Cog):
                 else None
             )
 
-        if "Invalid username" in str(error):
-            return (
-                await ctx.reply(
-                    embed=discord.Embed(
-                        title="Player not found",
-                        description="I could not find a ROBLOX player with that corresponding username.",
-                        color=BLANK_COLOR,
-                    )
-                )
-                if not do_not_send
-                else None
-            )
-
-        if isinstance(error, roblox.UserNotFound):
+        if isinstance(error, roblox.UserNotFound) or "Invalid username" in str(error):
             return (
                 await ctx.reply(
                     embed=discord.Embed(
@@ -274,7 +261,7 @@ class OnCommandError(commands.Cog):
                 await ctx.send(
                     embed=discord.Embed(
                         title="Missing Argument",
-                        description="You are missing a required argument to run this command.",
+                        description=f"You are missing a required argument to run this command.\n\n`{str(error).capitalize()}`",
                         color=BLANK_COLOR,
                     )
                 )
@@ -314,20 +301,55 @@ class OnCommandError(commands.Cog):
 
 
             if not do_not_send:
-                await ctx.send(
-                    embed=discord.Embed(
-                        title=f"{self.bot.emoji_controller.get_emoji('error')} Command Failure",
-                        description="The command you were attempting to run failed.\nContact ERM Support for assistance.",
-                        color=RED_COLOR,
-                    ).add_field(name="Error ID", value=f"[`{error_id}`]({config('SENTRY_BASE_URL') + error_link})", inline=False),
-                    view=View().add_item(
-                        Button(
-                            label="Contact ERM Support",
-                            style=discord.ButtonStyle.link,
-                            url="https://discord.gg/FAC629TzBy",
+                view = discord.ui.Container(accent_color=RED_COLOR)
+                view.add_item(
+                    discord.ui.TextDisplay(
+                        (
+                            f"### {self.bot.emoji_controller.get_emoji('error')} Command Failure\n"
+                            "A critical error has occured with ERM and the command has been stopped to prevent damage. Please contact ERM support for assistance and send them the error ID below.\n\n"
+                            f"**Error ID**\n`{error_id}`"
                         )
-                    ),
+                    )
+                ).add_item(discord.ui.Separator())
+                actionrow = discord.ui.ActionRow(
+                    discord.ui.Button(label = "Contact ERM Support", url="https://discord.gg/uAfU26VRa8"),
+                    discord.ui.Button(label = f"Error ID: {error_id}", disabled=True)
                 )
+                view.add_item(actionrow)
+                await ctx.send(
+                    view=discord.ui.LayoutView().add_item(view)
+                )
+    @commands.Cog.listener("on_error")
+    async def on_error(self, error):
+        bot = self.bot
+        error_id = error_gen()
+
+        if isinstance(error, discord.Forbidden):
+            if "Cannot send messages to this user" in str(error):
+                return
+
+        if isinstance(error, commands.CommandNotFound):
+            return
+        if isinstance(error, commands.CheckFailure):
+            return
+        if isinstance(error, commands.MissingRequiredArgument):
+            return
+        # # print(error)
+        # # print(str(error))
+        with push_scope() as scope:
+            scope.set_tag("error_id", error_id)
+            scope.level = "error"
+            await bot.errors.insert(
+                {
+                    "_id": error_id,
+                    "error": str(error),
+                    "time": datetime.datetime.now(tz=pytz.UTC).strftime(
+                        "%m/%d/%Y, %H:%M:%S"
+                    ),
+                }
+            )
+
+            capture_exception(error)
 
 async def setup(bot):
     await bot.add_cog(OnCommandError(bot))
